Reorganizing things into tool folders. Also, saving log10nbar for numerical precision
